Move main plugin function into its own file
