+++ src/main.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * The main plugin function
+ *
+ * @package create-wordpress-plugin
+ */
+
+namespace Create_WordPress_Plugin;
+
+use Alley\WP\Features\Group;
+
+/**
+ * Instantiate the plugin.
+ */
+function main(): void {
+	// Add features here.
+	$plugin = new Group();
+
+	$plugin->boot();
+}
+
